improvement to inference eval script

+++ tests/LiquidInf/Paper/Eval/eval.py
@@ -9,46 +9,49 @@ import time
 exe_name = str(subprocess.run(["cabal", "exec", "which", "Inference"], capture_output = True).stdout.decode('utf-8')).strip()
 
 def run_infer(file, name, timeout, extra_opts=[]):
-    # get info about the file
-    (funcs, depth) = get_counts(file)
+    if not os.path.exists("logs/" + name + ".txt"):
+        # get info about the file
+        (funcs, depth) = get_counts(file)
 
-    # actual run the test
-    # start_time = time.perf_counter();
-    # res = call_infer_process(file, timeout);
-    # end_time = time.perf_counter();
-    # elapsed = end_time - start_time;
-    (check_safe, res, res_err, counts, elapsed) = call_with_timing(file, timeout, extra_opts)
+        # actual run the test
+        # start_time = time.perf_counter();
+        # res = call_infer_process(file, timeout);
+        # end_time = time.perf_counter();
+        # elapsed = end_time - start_time;
+        (check_safe, res, res_err, counts, elapsed) = call_with_timing(file, timeout, extra_opts)
 
-    f = open("logs/" + name + ".txt", "w")
-    f.write(res.decode("utf-8") + "\n" + res_err.decode("utf-8"));
-    f.close();
+        f = open("logs/" + name + ".txt", "w")
+        f.write(res.decode("utf-8") + "\n" + res_err.decode("utf-8"));
+        f.close();
 
-    # MAKES EVERYTHING AFTER THIS TIMEOUT QUICKLY
-    timeout = "1";
+        # MAKES EVERYTHING AFTER THIS TIMEOUT QUICKLY
+        timeout = "1";
 
-    # run the test without extra fc
-    # no_fc_start_time = time.perf_counter();
-    # no_fc_res = call_infer_process(file, timeout, extra_opts + ["--no-use-extra-fc"])
-    # no_fc_end_time = time.perf_counter();
-    # no_fc_elapsed = no_fc_end_time - no_fc_start_time;
-    # no_fc_check_safe = no_fc_res.splitlines()[-2].decode('utf-8');
-    # no_fc_counts = get_opt_counts(no_fc_res);
-    (_, _, _, no_fc_counts, no_fc_elapsed) = call_with_timing(file, timeout, extra_opts + ["--no-use-extra-fc"])
+        # run the test without extra fc
+        # no_fc_start_time = time.perf_counter();
+        # no_fc_res = call_infer_process(file, timeout, extra_opts + ["--no-use-extra-fc"])
+        # no_fc_end_time = time.perf_counter();
+        # no_fc_elapsed = no_fc_end_time - no_fc_start_time;
+        # no_fc_check_safe = no_fc_res.splitlines()[-2].decode('utf-8');
+        # no_fc_counts = get_opt_counts(no_fc_res);
+        (_, _, _, no_fc_counts, no_fc_elapsed) = call_with_timing(file, timeout, extra_opts + ["--no-use-extra-fc"])
 
-    no_lev_dec_counts = empty_counts()
-    no_lev_dec_elapsed = None
-    if counts["searched_below"] is not None and int(counts["searched_below"]) > 0:
-        (_, _, _, no_lev_dec_counts, no_lev_dec_elapsed) = call_with_timing(file, timeout, extra_opts + ["--no-use-level-dec"])
+        no_lev_dec_counts = empty_counts()
+        no_lev_dec_elapsed = None
+        if counts["searched_below"] is not None and int(counts["searched_below"]) > 0:
+            (_, _, _, no_lev_dec_counts, no_lev_dec_elapsed) = call_with_timing(file, timeout, extra_opts + ["--no-use-level-dec"])
 
-    no_n_mdl_counts = empty_counts()
-    no_n_mdl_elapsed = None
-    if counts["negated_model"] is not None and int(counts["negated_model"]) > 0:
-        (_, _, _, no_n_mdl_counts, no_n_mdl_elapsed) = call_with_timing(file, timeout, extra_opts + ["--no-use-negated-models"])
+        no_n_mdl_counts = empty_counts()
+        no_n_mdl_elapsed = None
+        if counts["negated_model"] is not None and int(counts["negated_model"]) > 0:
+            (_, _, _, no_n_mdl_counts, no_n_mdl_elapsed) = call_with_timing(file, timeout, extra_opts + ["--no-use-negated-models"])
 
-    return (check_safe, elapsed, funcs, depth, counts
-                      , no_fc_elapsed, no_fc_counts
-                      , no_lev_dec_elapsed, no_lev_dec_counts
-                      , no_n_mdl_elapsed, no_n_mdl_counts);
+        return (check_safe, elapsed, funcs, depth, counts
+                          , no_fc_elapsed, no_fc_counts
+                          , no_lev_dec_elapsed, no_lev_dec_counts
+                          , no_n_mdl_elapsed, no_n_mdl_counts);
+    else:
+        None
 
 def call_with_timing(file, timeout, passed_args = []):
     start_time = time.perf_counter();
@@ -150,24 +153,26 @@ def test_pos_folder(folder, timeout, extra_opts=[]):
         if file.endswith(".lhs") or file.endswith(".hs"):
             print(file);
 
-            (check_safe, elapsed, funcs, depth, counts
-                       , no_fc_elapsed, no_fc_counts
-                       , no_lev_dec_elapsed, no_lev_dex_count
-                       , no_n_mdl_elapsed, no_n_mdl_count) = run_infer(os.path.join(folder, file), file, timeout, extra_opts);
+            res = run_infer(os.path.join(folder, file), file, timeout, extra_opts);
+            if res is not None:
+                (check_safe, elapsed, funcs, depth, counts
+                           , no_fc_elapsed, no_fc_counts
+                           , no_lev_dec_elapsed, no_lev_dex_count
+                           , no_n_mdl_elapsed, no_n_mdl_count) = res
 
-            if check_safe == "Safe":
-                print("\tSafe - " + str(elapsed) + "s");
-                safe_num += 1
-            elif check_safe == "Timeout":
-                print("\tTimeout")
-            else:
-                # print("check_safe =" + repr(check_safe) + "|")
-                print("\tUnsafe")
+                if check_safe == "Safe":
+                    print("\tSafe - " + str(elapsed) + "s");
+                    safe_num += 1
+                elif check_safe == "Timeout":
+                    print("\tTimeout")
+                else:
+                    # print("check_safe =" + repr(check_safe) + "|")
+                    print("\tUnsafe")
 
-            log.append((file, elapsed, funcs, depth, counts
-                            , no_fc_elapsed, no_fc_counts
-                            , no_lev_dec_elapsed, no_lev_dex_count
-                            , no_n_mdl_elapsed, no_n_mdl_count))
+                log.append((file, elapsed, funcs, depth, counts
+                                , no_fc_elapsed, no_fc_counts
+                                , no_lev_dec_elapsed, no_lev_dex_count
+                                , no_n_mdl_elapsed, no_n_mdl_count))
 
     return (log, safe_num, num_files)
